Setup rtl support.
