+++ library/src/main/java/com/aminography/primedatepicker/picker/go/MyNumberPicker.kt
@@ -1,12 +1,16 @@
 package com.aminography.primedatepicker.picker.go
 
 import android.content.Context
+import android.content.res.Resources
 import android.graphics.Color
 import android.graphics.Typeface
+import android.graphics.drawable.ColorDrawable
+import android.os.Build
 import android.util.AttributeSet
 import android.util.TypedValue
 import android.view.View
 import android.view.ViewGroup
+import android.widget.EditText
 import android.widget.NumberPicker
 import android.widget.TextView
 import com.aminography.primedatepicker.R
@@ -49,4 +53,34 @@ internal class MyNumberPicker : NumberPicker {
         internal var typeface: Typeface? = null
     }
 
+}
+
+internal fun NumberPicker.setDividerColor(color: Int) {
+    if (Build.VERSION.SDK_INT < Build.VERSION_CODES.P) {
+        val pickerFields = NumberPicker::class.java.declaredFields
+        for (pf in pickerFields) {
+            if (pf.name == "mSelectionDivider") {
+                pf.isAccessible = true
+                try {
+                    val colorDrawable = ColorDrawable(color)
+                    pf.set(this, colorDrawable)
+                } catch (e: IllegalArgumentException) {
+                    e.printStackTrace()
+                } catch (e: Resources.NotFoundException) {
+                    e.printStackTrace()
+                } catch (e: IllegalAccessException) {
+                    e.printStackTrace()
+                }
+                break
+            }
+        }
+    } else {
+        // TODO
+    }
+}
+
+internal fun NumberPicker.fixInputFilter() {
+    val field = NumberPicker::class.java.getDeclaredField("mInputText")
+    field.isAccessible = true
+    (field.get(this) as EditText).filters = arrayOfNulls(0)
 }
