+++ library/src/main/res/layout/range_days_header_rtl.xml
@@ -11,8 +11,8 @@
         android:background="@drawable/range_background_selector"
         app:layout_constrainedHeight="true"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@id/rangeEndBackView"
+        app:layout_constraintLeft_toRightOf="@id/rangeEndBackView"
+        app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
     <View
@@ -22,8 +22,8 @@
         android:background="@drawable/range_background_selector"
         app:layout_constrainedHeight="true"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@id/rangeStartBackView"
-        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toLeftOf="@id/rangeStartBackView"
         app:layout_constraintTop_toTopOf="parent" />
 
     <com.aminography.primedatepicker.picker.header.TwoLineTextView
